+++ src/main/java/domain/result/DealerProfit.java
@@ -1,3 +1,3 @@
 package domain.result;
 
-public record DealerProfit(int profit) {}
+public record DealerProfit(double profit) {}
