+++ src/main/java/strategy/BettingRule.java
@@ -4,5 +4,5 @@ import domain.result.GameResult;
 import domain.result.RoundBetInfo;
 
 public interface BettingRule {
-    int calculateBetAmount(RoundBetInfo roundBetInfo, GameResult gameResult);
+    double calculateBetAmount(RoundBetInfo roundBetInfo, GameResult gameResult);
 }
